# docs/security/openapi.yaml
openapi: 3.0.3
info:
  title: Civis CRM/ERP API
  version: 0.1.0
  description: |
    Generated from the current Next.js App Router endpoints under `app/api/**/route.ts`.
    This spec is intended for Swagger/SwaggerHub documentation and VAPT planning.
servers:
  - url: https://ik-crm-erp.onrender.com
    description: Production
  - url: http://localhost:3000
    description: Local development
...
components:
  securitySchemes:
    nextAuthSession:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth session cookie (development/non-secure cookie name).
    nextAuthSecureSession:
      type: apiKey
      in: cookie
      name: __Secure-next-auth.session-token
      description: NextAuth session cookie (secure cookie name).
  schemas:
...
    TaskInput:
      type: object
      properties:
        title: { type: string }
        dueDate: { type: string, format: date-time }
        ownerId: { type: string }
        relatedType: { type: string }
        relatedId: { type: string }
      required: [title]
    TaskUpdateInput:
      type: object
      properties:
        id: { type: string }
        status: { type: string, enum: [OPEN, DONE] }
      required: [id]
...
security:
  - nextAuthSession: []
  - nextAuthSecureSession: []
paths:
...
  /api/settings:
    get:
      tags: [Admin]
      summary: Get org settings and users
      responses:
        '200': { description: Org + users }
        '403': { description: Admin required }
    patch:
      tags: [Admin]
      summary: Update organization settings
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                theme: { type: string }
                notifyEmail: { type: string, format: email }
      responses:
        '200': { description: Updated org }
        '403': { description: Admin required }
    post:
      tags: [Admin]
      summary: Invite/create user in org
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                email: { type: string, format: email }
                role: { type: string, enum: [USER, ADMIN, SUPER_ADMIN] }
              required: [name, email, role]
      responses:
        '200': { description: User created }
        '400': { description: Validation or duplicate email }
        '403': { description: Admin required }
...
  /api/reports/summary:
    get:
      tags: [Reports]
      summary: Get report summary dataset
      parameters:
        - name: type
          in: query
          required: true
          schema: { type: string, enum: [accounting, crm, vat] }
      responses:
        '200': { description: Report summary }
        '400': { description: Invalid report type }
...
  /api/tasks:
    get:
      tags: [Tasks]
      summary: List tasks
      responses:
        '200': { description: Tasks list }
        '401': { description: Authentication required }
    post:
      tags: [Tasks]
      summary: Create task
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TaskInput' }
      responses:
        '200': { description: Task created }
        '400': { description: Validation error }
        '401': { description: Authentication required }
    patch:
      tags: [Tasks]
      summary: Update task status
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TaskUpdateInput' }
      responses:
        '200': { description: Task updated }
        '404': { description: Task not found }
...